Refactor: Separate packages\files
