--- Kotlin/src/main/kotlin/com/gildedrose/Item.kt
@@ -1,10 +1,0 @@
-package com.gildedrose
-
-//Check git works
-//open class Item(var name: String, var sellIn: Int, var quality: Int) {
-//    override fun toString(): String {
-//        return this.name + ", " + this.sellIn + ", " + this.quality
-//    }
-//}
-
-data class Item(var name: String, var sellIn: Int, var quality: Int)
