+++ Kotlin/src/main/kotlin/com/gildedrose/data/model/Item.kt
@@ -1,0 +1,3 @@
+package com.gildedrose.data.model
+
+data class Item(var name: String, var sellIn: Int, var quality: Int)
